requirements.txt
aniso8601==8.0.0
backports.functools-lru-cache==1.5
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.0
enum34==1.1.6
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
google-api-core==1.16.0
google-auth==1.10.1
google-cloud-vision==0.41.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
gunicorn==19.10.0
idna==2.8
imageio==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
moviepy==1.0.0
mysql-connector==2.2.9
numpy==1.16.6
opencv-python==4.1.2.30
pandas==0.24.2
Pillow==6.0.0
proglog==0.1.9
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.0
PySide2==5.12.3
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rsa==4.0
shiboken2==5.12.3
six==1.14.0
subprocess32==3.5.3
tqdm==4.31.1
urllib3==1.25.7
Werkzeug==0.16.0